refactor: move css generation to sdk
No changes. Just moved generateCss from react-sdk to sdk.
