--- packages/react-sdk/src/css/index.ts
@@ -1,2 +1,0 @@
-export * from "./global-rules";
-export * from "./css";
