+++ packages/react-sdk/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./remix";
-export * from "./css/index";
 export * from "./components/components-utils";
 export * from "./embed-template";
 export * from "./props";
